refactor(daa): split from HathorManager into separate module
