+++ hathor/transaction/resources/create_tx.py
@@ -20,6 +20,7 @@ from twisted.web import resource
 from hathor.api_util import set_cors
 from hathor.cli.openapi_files.register import register_resource
 from hathor.crypto.util import decode_address
+from hathor.daa import minimum_tx_weight
 from hathor.exception import InvalidNewTransaction
 from hathor.transaction import Transaction, TxInput, TxOutput
 from hathor.transaction.scripts import create_output_script
@@ -89,7 +90,7 @@ class CreateTxResource(resource.Resource):
         for tx_input in fake_signed_tx.inputs:
             # conservative estimate of the input data size to estimate a valid weight
             tx_input.data = b'\0' * 107
-        tx.weight = self.manager.minimum_tx_weight(fake_signed_tx)
+        tx.weight = minimum_tx_weight(fake_signed_tx)
         tx.verify_unsigned_skip_pow()
 
         if tx.is_double_spending():
