+++ hathor/wallet/resources/nano_contracts/execute.py
@@ -22,6 +22,7 @@ from twisted.web import resource
 from hathor.api_util import get_missing_params_msg, render_options, set_cors
 from hathor.cli.openapi_files.register import register_resource
 from hathor.crypto.util import decode_address
+from hathor.daa import minimum_tx_weight
 from hathor.transaction import Transaction, TxInput, TxOutput
 from hathor.transaction.scripts import P2PKH, NanoContractMatchValues
 from hathor.wallet.exceptions import InvalidAddress
@@ -99,7 +100,7 @@ class NanoContractExecuteResource(resource.Resource):
 
         tx.parents = self.manager.get_new_tx_parents()
         tx.update_timestamp(int(self.manager.reactor.seconds()))
-        tx.weight = self.manager.minimum_tx_weight(tx)
+        tx.weight = minimum_tx_weight(tx)
         tx.resolve()
         success = self.manager.propagate_tx(tx)
