+++ hathor/wallet/resources/send_tokens.py
@@ -22,6 +22,7 @@ from twisted.web.http import Request
 from hathor.api_util import render_options, set_cors
 from hathor.cli.openapi_files.register import register_resource
 from hathor.crypto.util import decode_address
+from hathor.daa import minimum_tx_weight
 from hathor.exception import InvalidNewTransaction
 from hathor.transaction import Transaction
 from hathor.transaction.exceptions import TxValidationError
@@ -125,7 +126,7 @@ class SendTokensResource(resource.Resource):
         tx.parents = values['parents']
         weight = values['weight']
         if weight is None:
-            weight = self.manager.minimum_tx_weight(tx)
+            weight = minimum_tx_weight(tx)
         tx.weight = weight
         tx.resolve()
         tx.verify()
