+++ tests/p2p/test_split_brain.py
@@ -3,8 +3,8 @@ import random
 import pytest
 from mnemonic import Mnemonic
 
+from hathor.daa import TestMode, _set_test_mode
 from hathor.graphviz import GraphvizVisualizer
-from hathor.manager import TestMode
 from hathor.simulator import FakeConnection
 from hathor.wallet import HDWallet
 from tests import unittest
@@ -31,8 +31,8 @@ class HathorSyncMethodsTestCase(unittest.TestCase):
         wallet = HDWallet(gap_limit=2)
         wallet._manually_initialize()
 
+        _set_test_mode(TestMode.TEST_ALL_WEIGHT)
         manager = super().create_peer(network, wallet=wallet)
-        manager.test_mode = TestMode.TEST_ALL_WEIGHT
         manager.avg_time_between_blocks = 64
 
         # Don't use it anywhere else. It is unsafe to generate mnemonic words like this.
