+++ tests/tx/test_blockchain.py
@@ -1,6 +1,7 @@
 from itertools import chain
 
 from hathor.conf import HathorSettings
+from hathor.daa import TestMode, _set_test_mode, get_weight_decay_amount
 from hathor.transaction import sum_weights
 from hathor.transaction.storage import TransactionMemoryStorage
 from tests import unittest
@@ -384,8 +385,8 @@ class BlockchainTestCase(unittest.TestCase):
 
     def test_daa_sanity(self):
         # sanity test the DAA
+        _set_test_mode(TestMode.DISABLED)
         manager = self.create_peer('testnet', tx_storage=self.tx_storage)
-        manager.test_mode = 0
         N = settings.BLOCK_DIFFICULTY_N_BLOCKS
         T = settings.AVG_TIME_BETWEEN_BLOCKS
         manager.avg_time_between_blocks = T
@@ -411,35 +412,46 @@ class BlockchainTestCase(unittest.TestCase):
             self.assertLess(new_weight, base_weight)
 
     def test_daa_weight_decay_amount(self):
-        manager = self.create_peer('testnet', tx_storage=self.tx_storage)
-        manager.test_mode = 0
+        _set_test_mode(TestMode.DISABLED)
         amount = settings.WEIGHT_DECAY_AMOUNT
 
         for distance in range(0, settings.WEIGHT_DECAY_ACTIVATE_DISTANCE, 10):
-            self.assertEqual(manager.get_weight_decay_amount(distance), 0)
+            self.assertEqual(get_weight_decay_amount(distance), 0)
 
         distance = settings.WEIGHT_DECAY_ACTIVATE_DISTANCE - 1
-        self.assertAlmostEqual(manager.get_weight_decay_amount(distance), 0)
+        self.assertAlmostEqual(get_weight_decay_amount(distance), 0)
 
         distance = settings.WEIGHT_DECAY_ACTIVATE_DISTANCE
         for k in range(1, 11):
             for _ in range(settings.WEIGHT_DECAY_WINDOW_SIZE):
-                self.assertAlmostEqual(manager.get_weight_decay_amount(distance), k * amount)
+                self.assertAlmostEqual(get_weight_decay_amount(distance), k * amount)
                 distance += 1
-        self.assertAlmostEqual(manager.get_weight_decay_amount(distance), 11 * amount)
+        self.assertAlmostEqual(get_weight_decay_amount(distance), 11 * amount)
 
     def test_daa_weight_decay_blocks(self):
+        from hathor import daa
+        orig_avg_time_between_blocks = daa.AVG_TIME_BETWEEN_BLOCKS
+        orig_min_block_weight = daa.MIN_BLOCK_WEIGHT
+
+        try:
+            self._test_daa_weight_decay_blocks()
+        finally:
+            daa.AVG_TIME_BETWEEN_BLOCKS = orig_avg_time_between_blocks
+            daa.MIN_BLOCK_WEIGHT = orig_min_block_weight
+
+    def _test_daa_weight_decay_blocks(self):
+        _set_test_mode(TestMode.DISABLED)
         manager = self.create_peer('testnet', tx_storage=self.tx_storage)
-        manager.test_mode = 0
         amount = settings.WEIGHT_DECAY_AMOUNT
 
-        manager.avg_time_between_blocks = settings.AVG_TIME_BETWEEN_BLOCKS
-        manager.min_block_weight = 2 + 2 * settings.WEIGHT_DECAY_AMOUNT
+        from hathor import daa
+        daa.AVG_TIME_BETWEEN_BLOCKS = settings.AVG_TIME_BETWEEN_BLOCKS
+        daa.MIN_BLOCK_WEIGHT = 2 + 2 * settings.WEIGHT_DECAY_AMOUNT
         add_new_blocks(manager, 2 * settings.BLOCK_DIFFICULTY_N_BLOCKS, advance_clock=settings.AVG_TIME_BETWEEN_BLOCKS)
-        manager.min_block_weight = 1
 
+        daa.MIN_BLOCK_WEIGHT = 1
         base_weight = manager.generate_mining_block().weight
-        self.assertGreater(base_weight, manager.min_block_weight)
+        self.assertGreater(base_weight, daa.MIN_BLOCK_WEIGHT)
 
         add_new_blocks(manager, 20, advance_clock=settings.AVG_TIME_BETWEEN_BLOCKS)
 
@@ -466,7 +478,7 @@ class BlockchainTestCase(unittest.TestCase):
 
         manager.reactor.advance(1)
         weight = manager.generate_mining_block().weight
-        self.assertAlmostEqual(weight, manager.min_block_weight)
+        self.assertAlmostEqual(weight, daa.MIN_BLOCK_WEIGHT)
 
 
 if __name__ == '__main__':
