+++ tests/tx/test_genesis.py
@@ -1,5 +1,5 @@
 from hathor.conf import HathorSettings
-from hathor.manager import TestMode
+from hathor.daa import TestMode, _set_test_mode, calculate_block_difficulty, minimum_tx_weight
 from hathor.transaction.storage import TransactionMemoryStorage
 from tests import unittest
 
@@ -62,13 +62,12 @@ class GenesisTest(unittest.TestCase):
         genesis_txs = [tx for tx in genesis if not tx.is_block]
         genesis_tx = genesis_txs[0]
 
-        network = 'testnet'
-        manager = self.create_peer(network, unlock_wallet=True)
-
         # Validate the block and tx weight
         # in test mode weight is always 1
-        self.assertEqual(manager.calculate_block_difficulty(genesis_block), 1)
-        self.assertEqual(manager.minimum_tx_weight(genesis_tx), 1)
-        manager.test_mode = TestMode.DISABLED
-        self.assertEqual(manager.calculate_block_difficulty(genesis_block), genesis_block.weight)
-        self.assertEqual(manager.minimum_tx_weight(genesis_tx), genesis_tx.weight)
+        _set_test_mode(TestMode.TEST_ALL_WEIGHT)
+        self.assertEqual(calculate_block_difficulty(genesis_block), 1)
+        self.assertEqual(minimum_tx_weight(genesis_tx), 1)
+
+        _set_test_mode(TestMode.DISABLED)
+        self.assertEqual(calculate_block_difficulty(genesis_block), genesis_block.weight)
+        self.assertEqual(minimum_tx_weight(genesis_tx), genesis_tx.weight)
